requirements.txt
absl-py==2.3.0
accelerate==1.4.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
airportsdata==20250523
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
binpacking==1.5.2
bitsandbytes==0.46.0
blake3==1.0.5
bleach==6.2.0
bypy==1.8.9
cachetools==6.0.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1749972191589/work/certifi
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.9.2
contourpy==1.3.2
cpm-kernels==1.0.11
crcmod==1.7
cryptography==45.0.4
cupy-cuda12x==13.4.1
cycler==0.12.1
dacite==1.9.2
datasets==3.3.2
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.4
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flash-attn==2.7.2.post1
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.12.0
-e git+ssh://git@github.com/juanjuanjuanbing/VLAD_eval.git@72be2e5b8195bb068393725261aa3155232b9ac1#egg=FT
future==1.0.0
gguf==0.10.0
googleapis-common-protos==1.70.0
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.2
icecream==2.1.4
idna==3.10
ijson==3.4.0
importlib_metadata==8.6.1
interegular==0.3.3
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
-e git+https://github.com/hiyouga/LLaMA-Factory.git@5308424705646e701da3ccae024bc05b76845a4b#egg=llamafactory
llguidance==0.7.25
llvmlite==0.43.0
lm-format-enforcer==0.10.11
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.6
mistune==3.1.3
modelscope==1.27.0
mpmath==1.3.0
-e git+ssh://git@github.com/modelscope/ms-swift.git@7c5be95482c96963beb47a26b31aa32cbcc9b499#egg=ms_swift
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.82.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
oss2==2.19.1
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip-chill==1.0.3
pipreqs==0.5.0
platformdirs==4.3.8
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.25.1
ruff==0.11.11
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.7
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
TorchCodec==0.2.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.18.2
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tyro==0.8.14
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
utils==1.0.2
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.8.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
